Generic * on OutOfMemory: clear MovingAverage counters * Alarm on OsLOadAverage -- added a chart that shows what happened in the last hour * Alarm on AlarmEventOsSwapping -- added a chart that shows what happened in the last hour * PersistWriterToHttpJson -- Added the ability to send specific Counter MOdels at a finer grain (for example so we can send: CmActiveStatements) * Upgrade H2 Database to 2.1.212
AseTune
* Alarm on AlarmEvent: ConfigResourceIsLow, ProcedureCacheLowOnMemory, ConfigResourceIsUsedUp -- added a chart that shows what happened in the last hour

SqlServerTune
* Fixed small buf when converting from sysdatabases to sys.databases (adoption to Azure databases)
* CmPerfCounters: Added Graph/Chart -- Target vs Total Memory
* DBMS Config -- Added check to check if 'instant file initialization' is enabled at the OS Level

+++ src/com/asetune/pcs/H2DatabaseEventListener.java
@@ -148,7 +148,8 @@ public class H2DatabaseEventListener implements DatabaseEventListener
 	 * @param max the highest value
 	 */
 	@Override
-	public void setProgress(int state, String name, int x, int max)
+//	public void setProgress(int state, String name, int x, int max)
+	public void setProgress(int state, String name, long x, long max)
 	{
 //		int STATE_SCAN_FILE          = 0; // This state is used when scanning the database file.
 //		int STATE_CREATE_INDEX       = 1; // This state is used when re-creating an index.
